requirements.txt
sanic
sanic-swagger
motor
pymongo
python-dotenv
aiohttp
web3
pydantic
